move common parts to base class
