+++ vocode/streaming/audio/audio_service.py
@@ -1,6 +1,5 @@
 import logging
 from typing import Optional
-import queue
 
 from vocode.streaming.models.audio import AudioServiceConfig
 from vocode.streaming.audio.base_audio_service import BaseThreadAsyncAudioService
@@ -14,13 +13,10 @@ class AudioService(BaseThreadAsyncAudioService[AudioServiceConfig]):
         audio_service_config: AudioServiceConfig,
         logger: Optional[logging.Logger] = None,
     ):
-        super().__init__(audio_service_config)
-        self.logger = logger
+        super().__init__(audio_service_config, logger)
 
-        self._ended = False
-        self.is_ready = False
-
-    def process(self, chunk):
+    def process(self, chunk: bytes) -> bytes:
+        """No processing"""
         return chunk
 
     def _run_loop(self):
@@ -32,31 +28,3 @@ class AudioService(BaseThreadAsyncAudioService[AudioServiceConfig]):
 
             if self._ended:
                 break
-
-    def generator(self):
-        """audio frame generator"""
-        while not self._ended:
-            try:
-                chunk = self.input_janus_queue.sync_q.get(timeout=5)
-            except queue.Empty:
-                return
-
-            if chunk is None:
-                return
-
-            data = [chunk]
-            # Now consume whatever other data's still buffered.
-            while True:
-                try:
-                    chunk = self.input_janus_queue.sync_q.get_nowait()
-                    if chunk is None:
-                        return
-                    data.append(chunk)
-                except queue.Empty:
-                    break
-
-            yield b"".join(data)
-
-    def terminate(self):
-        self._ended = True
-        super().terminate()
